functions/requirements.txt
# Function dependencies, for example:
# package>=version

langchain==0.0.198
pinecone-client==2.2.1
google-generativeai==0.1.0
sentence_transformers==2.2.2
google-cloud-storage>=1.32.0
google-cloud-aiplatform>=1.25.0
firebase-admin
pdfminer.six